+++ src/Plugin/AssetEntry.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const Options = require('./Options');
+const PluginService = require('./PluginService');
 const { isFunction } = require('../Common/Helpers');
 
 /**
@@ -180,6 +181,10 @@ class AssetEntry {
     // skip duplicate entries
     if (this.#exists(name, importFile)) return false;
 
+    const compilation = this.compilation;
+    const EntryPlugin = compilation.compiler.webpack.EntryPlugin;
+    const entryDependency = EntryPlugin.createDependency(importFile, { name });
+
     const entry = {
       name,
       runtime: undefined,
@@ -210,13 +215,15 @@ class AssetEntry {
     this.#add(entry, assetEntryOptions);
     this.compilationEntryNames.add(name);
 
-    // adds the entry of the script from the template to the compilation
-    // see reference: node_modules/webpack/lib/EntryPlugin.js
-    const entryDependency = this.EntryPlugin.createDependency(importFile, { name });
-
-    this.compilation.addEntry(context, entryDependency, entry, (err) => {
+    // add the script resolved in the template to the compilation
+    compilation.addEntry(context, entryDependency, entry, (err) => {
       if (err) throw new Error(err);
     });
+
+    // add missing dependencies after rebuild
+    if (PluginService.isWatchMode()) {
+      new EntryPlugin(context, importFile, { name }).apply(compilation.compiler);
+    }
 
     return true;
   }