+++ test/manual/watch-dependencies-import-module/src/views/app.js
@@ -1,0 +1,3 @@
+import 'bootstrap';
+
+console.log('>> To test change me: 123');